requirements.txt
geopy==2.2.0
googlemaps==4.5.3
pandas==1.3.5
protobuf==3.19.1
PyYAML==6.0
google-cloud-storage==1.43.0